civiclens-ai/ai-microservice/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
openai==1.3.7
requests==2.31.0
python-dotenv==1.0.0
deep-translator==1.11.4
googlemaps==4.10.0
whisper==1.1.10
torch==2.1.1
torchaudio==2.1.1
librosa==0.10.1
numpy==1.24.3
scipy==1.11.4
aiofiles==23.2.1
httpx==0.25.2
tenacity==8.2.3
langdetect==1.0.9
pydub==0.25.1
ffmpeg-python==0.2.0